+++ doc/Changelog.md
@@ -1,6 +1,7 @@
 * mark media as selected when coming from single media
 * mark subalbum as selected when coming from it
 * make page scroll to subalbum when going up in album tree
+* make selected subalbum/thumbnail visible when resizing
 
 ### version v4.9.46 (Apr 23, 2021)